requirements.txt
numpy==1.22.4
torch==2.2.2
matplotlib==3.6.2
pandas==2.0.1
scikit_learn==1.1.3
pykan
pyyaml
tqdm==4.66.2
setuptools==65.5.0
sympy==1.11.1
seaborn
scanpy
